+++ Sources/MapboxDirections/RouteResponse.swift
@@ -10,7 +10,7 @@ public struct RouteResponse {
     
     public let identifier: String?
     public var routes: [Route]?    
-    public let waypoints: [Waypoint]?
+    public let waypoints: [Route.Waypoint]?
     
     public let options: ResponseOptions
     public let credentials: DirectionsCredentials
@@ -27,15 +27,12 @@ public struct RouteResponse {
 
 extension RouteResponse: Codable {
     enum CodingKeys: String, CodingKey {
-        case code
-        case message
-        case error
         case identifier = "uuid"
         case routes
         case waypoints
     }
     
-    public init(httpResponse: HTTPURLResponse?, identifier: String? = nil, routes: [Route]? = nil, waypoints: [Waypoint]? = nil, options: ResponseOptions, credentials: DirectionsCredentials) {
+    public init(httpResponse: HTTPURLResponse?, identifier: String? = nil, routes: [Route]? = nil, waypoints: [Route.Waypoint]? = nil, options: ResponseOptions, credentials: DirectionsCredentials) {
         self.httpResponse = httpResponse
         self.identifier = identifier
         self.routes = routes
@@ -58,12 +55,12 @@ extension RouteResponse: Codable {
             routes = try decoder.decode([Route].self, from: matchesData)
         }
         
-        var waypoints: [Waypoint]?
+        var waypoints: [Route.Waypoint]?
         
         if let tracepoints = response.tracepoints {
             let filtered = tracepoints.compactMap { $0 }
             let tracepointsData = try encoder.encode(filtered)
-            waypoints = try decoder.decode([Waypoint].self, from: tracepointsData)
+            waypoints = try decoder.decode([Route.Waypoint].self, from: tracepointsData)
         }
     
         self.init(httpResponse: response.httpResponse, identifier: nil, routes: routes, waypoints: waypoints, options: .match(options), credentials: credentials)
@@ -91,37 +88,63 @@ extension RouteResponse: Codable {
         self.identifier = try container.decodeIfPresent(String.self, forKey: .identifier)
         
         // Decode waypoints from the response and update their names according to the waypoints from DirectionsOptions.waypoints.
-        let decodedWaypoints = try container.decodeIfPresent([Waypoint?].self, forKey: .waypoints)?.compactMap{ $0 }
-        var optionsWaypoints: [Waypoint] = []
+        let decodedWaypoints = try container.decodeIfPresent([Route.Waypoint?].self, forKey: .waypoints)?.compactMap{ $0 }
+        var optionsWaypoints: [Route.Waypoint] = []
         
         switch options {
         case let .match(options: matchOpts):
-            optionsWaypoints = matchOpts.waypoints
+            optionsWaypoints = matchOpts.waypoints.map {
+                Route.Waypoint(coordinate: $0.coordinate,
+                               correction: 0,
+                               name: $0.name)
+            }
         case let .route(options: routeOpts):
-            optionsWaypoints = routeOpts.waypoints
+            optionsWaypoints = routeOpts.waypoints.map {
+                Route.Waypoint(coordinate: $0.coordinate,
+                               correction: 0,
+                               name: $0.name)
+            }
         }
                 
         if let decodedWaypoints = decodedWaypoints {
             // The response lists the same number of tracepoints as the waypoints in the request, whether or not a given waypoint is leg-separating.
-            waypoints = zip(decodedWaypoints, optionsWaypoints).map { (pair) -> Waypoint in
-                let (decodedWaypoint, waypointInOptions) = pair
-                let waypoint = Waypoint(coordinate: decodedWaypoint.coordinate, coordinateAccuracy: waypointInOptions.coordinateAccuracy, name: waypointInOptions.name?.nonEmptyString ?? decodedWaypoint.name)
-                waypoint.separatesLegs = waypointInOptions.separatesLegs
-                return waypoint
+            waypoints = zip(decodedWaypoints, optionsWaypoints).map { (pair) -> Route.Waypoint in
+                var (decodedWaypoint, waypointInOptions) = pair
+                if let name = waypointInOptions.name?.nonEmptyString {
+                    decodedWaypoint.name = name
+                }
+                return decodedWaypoint
             }
-            waypoints?.first?.separatesLegs = true
-            waypoints?.last?.separatesLegs = true
         } else {
             waypoints = decodedWaypoints
         }
         
         if let routes = try container.decodeIfPresent([Route].self, forKey: .routes) {
             // Postprocess each route.
+            var legSeparators = waypoints?.compactMap { $0 }
+            if let options = decoder.userInfo[.options] as? DirectionsOptions, let waypoints = waypoints {
+                // Select first, last and all other waypoints which separate legs.
+                var optionsLegSeparators = options.legSeparators
+                legSeparators = zip(waypoints, options.waypoints).compactMap { (pair) -> Route.Waypoint? in
+                    let (decodedWaypoint, waypointInOptions) = pair
+                    if decodedWaypoint == waypoints.last! ||
+                        decodedWaypoint == waypoints.first! {
+                        return decodedWaypoint
+                    }
+                    
+                    guard let index = optionsLegSeparators.firstIndex(of: waypointInOptions) else {
+                        return nil
+                    }
+                    optionsLegSeparators.remove(at: index)
+                    return decodedWaypoint
+                }
+            }
+            
             for route in routes {
                 route.routeIdentifier = identifier
                 // Imbue each route’s legs with the waypoints refined above.
-                if let waypoints = waypoints {
-                    route.legSeparators = waypoints.filter { $0.separatesLegs }
+                if let legSeparators = legSeparators {
+                    route.legSeparators = legSeparators
                 }
             }
             self.routes = routes