+++ Tests/MapboxDirectionsTests/RoutableMatchTests.swift
@@ -28,12 +28,13 @@ class RoutableMatchTest: XCTestCase {
         
         var routeResponse: RouteResponse!
         
-        let matchOptions = MatchOptions(coordinates: locations)
+        let matchOptions = MatchOptions(waypoints: locations.map { (location) in
+            var waypoint = MatchOptions.Waypoint(coordinate: location)
+            waypoint.separatesLegs = false
+            return waypoint
+        })
         matchOptions.includesSteps = true
         matchOptions.routeShapeResolution = .full
-        for waypoint in matchOptions.waypoints[1..<(locations.count - 1)] {
-            waypoint.separatesLegs = false
-        }
         
         let task = Directions(credentials: BogusCredentials).calculateRoutes(matching: matchOptions) { (session, result) in